requirements.txt
fastapi
uvicorn[standard]
gunicorn
google-cloud-texttospeech
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-genai==1.13.0
pydantic
starlette
itsdangerous
pypdf[crypto]
Jinja2